PyPy specific copy of _hashing.pyx and _svmlight_format.pyx
