+++ build_tools/travis/install.sh
@@ -18,12 +18,12 @@ echo 'pip:'
 ls $HOME/.cache/pip
 
 if [[ "$DISTRIB" == "pypy" ]]; then
+    ls ~/virtualenv/
     pip install numpy Cython Tempita
     # use scipy 1.1 when it's released
-    pip install git+https://github.com/scipy/scipy.git@b3162e187174fa0b1462f7b243c483285b38f45e
+    pip install -vv git+https://github.com/scipy/scipy.git@b3162e187174fa0b1462f7b243c483285b38f45e
     pip install -e .
 else 
-    python --version
     echo "skip non PyPy builds for now"
     exit 1
 fi
