+++ sklearn/utils/__init__.py
@@ -2,6 +2,7 @@
 The :mod:`sklearn.utils` module includes various utilities.
 """
 from collections import Sequence
+import sys
 
 import numpy as np
 from scipy.sparse import issparse
@@ -26,6 +27,9 @@ __all__ = ["murmurhash3_32", "as_float_array",
            "column_or_1d", "safe_indexing",
            "check_consistent_length", "check_X_y", 'indexable',
            "check_symmetric", "indices_to_mask", "deprecated"]
+
+# TODO: not sure this is the best place for this flag
+IS_PYPY = '__pypy__' in sys.modules
 
 
 class Bunch(dict):
